requirements.txt
# LiveKit Core Dependencies - PINNED TO WORKING VERSIONS
livekit-agents==1.2.1
livekit-plugins-openai==0.8.1
livekit-plugins-speechmatics==0.6.1
livekit-plugins-silero==0.6.1

# AI/ML Dependencies
openai>=1.0.0

# Web Framework (for health checks)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
aiohttp>=3.8.0

# Database
asyncpg>=0.29.0

# Environment & Configuration
python-dotenv>=1.0.0

# Logging & Monitoring
# structlog>=23.0.0  # Removed - using standard logging

# Production Dependencies
gunicorn>=21.0.0
psutil>=5.9.0

# Audio Processing
pyaudio>=0.2.11

# Async utilities
asyncio-throttle>=1.0.0